refactor: 모달 상태, 핸들러, 데이터 연동 로직 분리
